% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaffold_shinyApp.R
\name{scaffold_dashboard}
\alias{scaffold_dashboard}
\title{Scaffold Shiny Dashboards}
\usage{
scaffold_dashboard(root = NULL, app.name = NULL, delete.existing = FALSE,
  run = TRUE, title = "Test App", libraries = NULL, sources = NULL,
  load = NULL, modules = NULL)
}
\arguments{
\item{root}{path of the directory in which the folder with app files is to
be created.}

\item{app.name}{App name.}

\item{delete.existing}{If the folder with same name as `app.name` already
exists, should it be overwritten? Default is FALSE.}

\item{run}{Run the resulting app? Default is TRUE.}

\item{libraries}{}

\item{sources}{}

\item{load}{}

\item{modules}{}

\item{tilte}{}
}
\description{
This functions generates boilerplate code for a fully functioning
shiny-dashboard  applications.
}
\examples{
\dontrun{
scaffold_dashboard(root = '~/Downloads', app.name = 'testApp',
delete.existing = TRUE)

}
}
\author{
Janko Cizel
}

