% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaffold_module_plmRegression.R
\name{scaffold_module_plmRegression}
\alias{scaffold_module_plmRegression}
\title{instantShiny module: PLM Regression}
\usage{
scaffold_module_plmRegression(id = "<id>", data = "<data>",
  idVar = "<idVar>", dateVar = "<dateVar>", app.dir = ".")
}
\arguments{
\item{id}{ID char string to be appended to input and output shiny objects}

\item{data}{name of data file}

\item{idVar}{name of variable that defines cross-sectional dimension of the data}

\item{dateVar}{name of variable that defines time-dimension of the data}

\item{app.dir}{path of the directory in which to insert the module}
}
\description{
Function that scaffolds the instantShiny module
}
\author{
Janko Cizel
}

